requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# File handling
python-multipart==0.0.17

# LangGraph and LangChain
langgraph==0.2.45
langchain-core==0.3.15
langchain-google-genai==2.0.5

# Google AI
google-generativeai==0.8.3

# PDF processing
pymupdf==1.24.13

# Data validation and settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment variables
python-dotenv==1.0.1